Using the if statement with the method we already created
- Using the OwnedBy method in the Post model
- The request variable was added in the destroy method in the PostController
- Check the post ownedBy the authenticated user
- Do the same experiment again by inspecting the delete button
- The error message was added.

+++ app/Http/Controllers/PostController.php
@@ -34,9 +34,13 @@ class PostController extends Controller
 
     }
 
-    public function destroy(Post $post)
+    public function destroy(Request $request, Post $post)
     {
-        dd($post);
+
+        if (!$post->ownedBy($request->user())) {
+            return response(null, 402);
+        }
+
         $post->delete();
         return back();
     }
